requirements.txt
fairseq==0.12.2
h5py==3.7.0
hydra-core==1.0.7
intervaltree==3.1.0
iopath==0.1.10
iterative_stratification==0.1.7
librosa==0.10.1
matplotlib==3.6.3
numpy==1.23.5
omegaconf==2.0.6
packaging==24.0
pandas==1.5.3
pyarrow==15.0.2
scikit_learn==1.2.0
scipy==1.10.0
seaborn==0.13.2
scikit-image==0.23.1
soundfile==0.12.1
tensorflow==2.11.0
timm==0.6.12
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1